refactor: backoff 지수시간, random: true로 설정, 테스트 코드 수정
