+++ main/main/src/main/resources/static/docs/index.html
@@ -527,8 +527,8 @@ Host: localhost:8080
   "name" : "AmorGakCo",
   "description" : "모각코 합시다.",
   "groupCapacity" : 4,
-  "beginAt" : "2024-10-01T22:18:36.28155",
-  "endAt" : "2024-10-02T01:18:36.28155",
+  "beginAt" : "2024-10-02T00:24:59.33542",
+  "endAt" : "2024-10-02T03:24:59.33542",
   "latitude" : 37.5703901,
   "longitude" : 126.9748397,
   "address" : "서울특별시 종로구 신문로1가 23"
@@ -608,8 +608,8 @@ Content-Length: 359
   "data" : {
     "hostNickname" : "아모르겠고",
     "hostImgUrl" : "https://fakeimg",
-    "beginAt" : "2024-10-01T22:18:32.484159",
-    "endAt" : "2024-10-02T01:18:32.484169",
+    "beginAt" : "2024-10-02T00:24:58.038657",
+    "endAt" : "2024-10-02T03:24:58.038666",
     "groupCapacity" : 4,
     "currentParticipants" : 3,
     "address" : "서울특별시 종로구 신문로1가 23"
@@ -701,8 +701,8 @@ Content-Length: 880
     "address" : "서울특별시 종로구 신문로1가 23",
     "longitude" : 126.9748397,
     "latitude" : 37.5703901,
-    "beginAt" : "2024-10-01T22:18:32.484159",
-    "endAt" : "2024-10-02T01:18:32.484169",
+    "beginAt" : "2024-10-02T00:24:58.038657",
+    "endAt" : "2024-10-02T03:24:58.038666",
     "groupMembers" : [ {
       "memberId" : 2,
       "imgUrl" : "https://fakeimg",
@@ -1102,39 +1102,39 @@ Content-Length: 1489
       "groupId" : 4,
       "name" : "AmorGakCo",
       "address" : "서울특별시 종로구 신문로1가 23",
-      "beginAt" : "2024-10-01T22:18:32.514261",
-      "endAt" : "2024-10-02T01:18:32.514271"
+      "beginAt" : "2024-10-02T00:24:58.063981",
+      "endAt" : "2024-10-02T03:24:58.063987"
     }, {
       "groupId" : 5,
       "name" : "AmorGakCo",
       "address" : "서울특별시 종로구 신문로1가 23",
-      "beginAt" : "2024-10-01T22:18:32.514261",
-      "endAt" : "2024-10-02T01:18:32.514271"
+      "beginAt" : "2024-10-02T00:24:58.063981",
+      "endAt" : "2024-10-02T03:24:58.063987"
     }, {
       "groupId" : 6,
       "name" : "AmorGakCo",
       "address" : "서울특별시 종로구 신문로1가 23",
-      "beginAt" : "2024-10-01T22:18:32.514261",
-      "endAt" : "2024-10-02T01:18:32.514271"
+      "beginAt" : "2024-10-02T00:24:58.063981",
+      "endAt" : "2024-10-02T03:24:58.063987"
     } ],
     "InactivatedGroup" : [ {
       "groupId" : 1,
       "name" : "AmorGakCo",
       "address" : "서울특별시 종로구 신문로1가 23",
-      "beginAt" : "2024-10-01T22:18:32.514261",
-      "endAt" : "2024-10-02T01:18:32.514271"
+      "beginAt" : "2024-10-02T00:24:58.063981",
+      "endAt" : "2024-10-02T03:24:58.063987"
     }, {
       "groupId" : 2,
       "name" : "AmorGakCo",
       "address" : "서울특별시 종로구 신문로1가 23",
-      "beginAt" : "2024-10-01T22:18:32.514261",
-      "endAt" : "2024-10-02T01:18:32.514271"
+      "beginAt" : "2024-10-02T00:24:58.063981",
+      "endAt" : "2024-10-02T03:24:58.063987"
     }, {
       "groupId" : 3,
       "name" : "AmorGakCo",
       "address" : "서울특별시 종로구 신문로1가 23",
-      "beginAt" : "2024-10-01T22:18:32.514261",
-      "endAt" : "2024-10-02T01:18:32.514271"
+      "beginAt" : "2024-10-02T00:24:58.063981",
+      "endAt" : "2024-10-02T03:24:58.063987"
     } ]
   }
 }</code></pre>
